backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-community>=0.0.10
pypdf>=3.17.1
openai>=1.6.1
python-dotenv>=1.0.0
elasticsearch>=8.11.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
requests>=2.31.0
aiohttp>=3.9.1
httpx>=0.26.0
